backend/requirements.txt
Flask
Flask-SQLAlchemy
python-dotenv
Werkzeug
PyJWT
